refactor(test): standardize Behave common* contract modules
Refs #216\n\n- Promote shared helper contracts to common_diagnostic_utils/common_workspace\n- Remove legacy shim modules and update step imports\n- Document common* cross-repo contract convention in features docs\n\nCo-Authored-By: Oz <oz-agent@warp.dev>
